Introduce job system
The idea is to have a queue of jobs that will be consumed by a pool of
threads. These changes define Job, Queue and Counter.
Job is basically a pointer to the logic of the job, the parameters and a
reference to an atomic counter to decrement once the task is done.
Queue is a queue of jobs implemeted with mutexes.
Coutner is an atomic unsigned int that will be incremented on job
scheduling and decremented on job completition. Used by the game loop to
figure out when all the tasks of a frame have been completed.

Remaining work: Define and spawn thread workers.